chatbot_backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.5.2
google-cloud-aiplatform==1.49.0
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.26.0
google-cloud-storage==2.10.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.24.3
google-generativeai>=0.4.0
httpx==0.25.1
pillow==10.1.0
faiss-cpu==1.7.4
firebase-admin>=6.0.0,<7.0.0